+++ src/commands/nwdCommands/ls.js
